+++ Venmo/src/main/java/com/braintreepayments/api/VenmoRequest.java
@@ -25,7 +25,7 @@ public class VenmoRequest implements Parcelable {
     private String taxAmount;
     private String shippingAmount;
     private ArrayList<VenmoLineItem> lineItems;
-    private boolean finalAmount;
+    private boolean isFinalAmount;
 
     private final @VenmoPaymentMethodUsage int paymentMethodUsage;
 
@@ -250,18 +250,25 @@ public class VenmoRequest implements Parcelable {
     }
 
     /**
-     * @param finalAmount Optional - Indicates whether the purchase amount is the final amount.
+     * @param isFinalAmount Optional - Indicates whether the purchase amount is the final amount.
      *                    Defaults to false.
      */
-    public void setFinalAmount(boolean finalAmount) {
-        this.finalAmount = finalAmount;
+    public void setIsFinalAmount(boolean isFinalAmount) {
+        this.isFinalAmount = isFinalAmount;
+    }
+
+    /**
+     * @return The boolean value of the flag that signifies whether the purchase amount is the final amount.
+     */
+    public boolean getIsFinalAmount() {
+        return isFinalAmount;
     }
 
     /**
      * @return Whether or not the purchase amount is the final amount as a string value.
      */
-    String getFinalAmountAsString() {
-        return String.valueOf(this.finalAmount);
+    String getIsFinalAmountAsString() {
+        return String.valueOf(this.isFinalAmount);
     }
 
 
@@ -278,7 +285,7 @@ public class VenmoRequest implements Parcelable {
         taxAmount = in.readString();
         totalAmount = in.readString();
         lineItems = in.createTypedArrayList(VenmoLineItem.CREATOR);
-        finalAmount = in.readByte() != 0;
+        isFinalAmount = in.readByte() != 0;
     }
 
     public static final Creator<VenmoRequest> CREATOR = new Creator<VenmoRequest>() {
@@ -312,6 +319,6 @@ public class VenmoRequest implements Parcelable {
         parcel.writeString(taxAmount);
         parcel.writeString(totalAmount);
         parcel.writeTypedList(lineItems);
-        parcel.writeByte((byte) (finalAmount ? 1 : 0));
+        parcel.writeByte((byte) (isFinalAmount ? 1 : 0));
     }
 }